[ADD] Manejo de archivos

+++ CEDigital_api/Models/Sql/Documento.cs
@@ -7,7 +7,7 @@ namespace CEDigital_api.Models.Sql
         [Key]
         public int id_documento { get; set; }
         public string nombre_archivo { get; set; }
-        public float size { get; set; }
+        public double size { get; set; }
         public DateTime fecha_subida { get; set; }
 
         // Foreign Keys
